hw_9/pyproject.toml
[project]
name = "hw-9"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "selenium (>=4.32.0,<5.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "allure-pytest (>=2.14.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
]

[tool.poetry]
packages = [{include = "hw_9", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
